refactor(ChBottomSheet): move use-modal-bottom-sheet
Move use-modal-bottom-sheet-contoller to separate folder
