+++ src/composable/modal-bottom-sheet-controller/index.ts
@@ -1,0 +1,5 @@
+export { useModalBottomSheetController } from './use-modal-bottom-sheet-controller'
+export type {
+  ModalBottomSheetController,
+  ModalBottomSheet
+} from './use-modal-bottom-sheet-controller'
