+++ src/index.js
@@ -1,4 +1,4 @@
-export { default as ChButton } from './components/ChButton/'
+export { default as ChButton } from './components/ChButton'
 export { default as ChCheckbox } from './components/ChCheckbox'
 export { default as ChInput } from './components/ChInput'
 export { default as ChRadio } from './components/ChRadio'
@@ -9,5 +9,5 @@ export { default as ChToggleButton } from './components/ChToggleButton'
 export { default as ChToggleButtonGroup } from './components/ChToggleButtonGroup'
 export { default as ChTooltip } from './components/ChTooltip'
 export { default as ChBottomSheet } from './components/ChBottomSheet'
-export { useModalBottomSheetController } from './composable/use-modal-bottom-sheet-controller'
+export { useModalBottomSheetController } from './composable/modal-bottom-sheet-controller'
 export { default as ChSpinner } from './components/ChSpinner'
